renaming the Verifications.cs file to the VerifyExtensions.cs and moving this file to the `Utils` folder
