+++ ANTWebAPI/src/ANTWebAPI.BLL/UseCases/CatalogUseCases.cs
@@ -1,5 +1,6 @@
 ﻿using ANTWebAPI.BLL.Models;
 using ANTWebAPI.BLL.Repositories;
+using ANTWebAPI.BLL.Utils;
 
 namespace ANTWebAPI.BLL.UseCases;
